refactor(modals): merge BulkBuyModal and BulkSellModal into BulkTradeModal
